# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: SkillSnap - AI-powered local services marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers/{id}/reviews:
    get:
      operationId: getProviderReviews
      tags: [providers]
      summary: Get reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Provider reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
  /reviews:
    get:
      operationId: listMyReviews
      tags: [reviews]
      summary: List reviews written by the authenticated consumer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reviews for current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /payments:
    get:
      operationId: listMyPayments
      tags: [payments]
      summary: List payments for the authenticated consumer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRecord"
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate payment for a booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
...
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin]
      summary: Payments with consumer/provider names (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPaymentsPage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateReviewInput:
      type: object
      properties:
        bookingId:
          type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - bookingId
        - rating
        - comment
    Review:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        consumerId:
          type: string
        providerId:
          type: string
        rating:
          type: number
        comment:
          type: string
        consumerName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - consumerId
        - providerId
        - rating
        - comment
        - createdAt
...
    InitiatePaymentInput:
      type: object
      properties:
        bookingId:
          type: string
        amount:
          type: number
      required:
        - bookingId
        - amount
    PaymentResponse:
      type: object
      properties:
        paymentId:
          type: string
        status:
          type: string
        amount:
          type: number
        currency:
          type: string
        message:
          type: string
      required:
        - paymentId
        - status
        - amount
        - currency
        - message
    PaymentRecord:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        provider:
          type: string
        externalPaymentId:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        paidAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - amount
        - currency
        - status
        - createdAt
...
    AdminPaymentsPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: number
      required:
        - items
        - total